requirements_all.txt
accelerate==0.24.1
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
datasets==2.14.6
debugpy==1.8.0
decorator==5.1.1
deepspeed==0.12.2
defusedxml==0.7.1
diffusers==0.23.0
dill==0.3.7
exceptiongroup==1.1.3
executing==2.0.1
fastapi==0.104.1
fastjsonschema==2.18.1
ffmpy==0.3.1
filelock==3.9.0
flatbuffers==23.5.26
fonttools==4.44.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.4.0
ftfy==6.1.1
gitdb==4.0.11
GitPython==3.1.40
gradio==4.2.0
gradio_client==0.7.0
h11==0.14.0
hjson==3.1.0
httpcore==1.0.2
httpx==0.25.1
huggingface-hub==0.17.3
humanfriendly==10.0
idna==2.10
imageio==2.32.0
importlib-metadata==6.8.0
importlib-resources==6.1.1
ipykernel==6.26.0
ipython==8.16.1
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.0
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.10.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.1
kiwisolver==1.4.5
lazy_loader==0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.0
ninja==1.11.1.1
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.24.1
onnxruntime-gpu==1.16.2
open-clip-torch==2.23.0
opencv-python==4.8.1.78
orjson==3.9.10
overrides==7.4.0
packaging==23.2
pandas==2.1.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==4.0.0
prometheus-client==0.18.0
prompt-toolkit==3.0.40
protobuf==4.25.0
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==14.0.1
pycocotools==2.0.7
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydub==0.25.1
Pygments==2.16.1
pynvml==11.5.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.5.0
QtPy==2.4.1
referencing==0.30.2
regex==2023.10.3
requests==2.25.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.6.0
rpds-py==0.12.0
safetensors==0.4.0
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.3
semantic-version==2.10.0
Send2Trash==1.8.2
sentencepiece==0.1.99
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
starlette==0.27.0
sympy==1.12
terminado==0.18.0
threadpoolctl==3.2.0
tifffile==2023.9.26
timm==0.9.10
tinycss2==1.2.1
tokenizers==0.14.1
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.0
torch==2.1.0+cu118
torchattacks==3.5.1
torchaudio==2.1.0+cu118
torchvision==0.16.0+cu118
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
transformers==4.35.0
triton==2.1.0
typer==0.9.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.18
uvicorn==0.24.0.post1
wcwidth==0.2.9
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
websockets==11.0.3
widgetsnbextension==4.0.9
-e git+ssh://git@github.com/johnding1996/Watermark-Experiments.git@c0e7bb94123eade85450e7d4f8bab2a4bc0b3de1#egg=wmbench
xformers==0.0.22.post7+cu118
xxhash==3.4.1
yarl==1.9.2
zipp==3.17.0
